Make session look for credentials to download warcs
Not sure if code change is the best way to resolve the issue, but we might want to do something about this:

I've found that I can't use this script to successfully download WARCs without this change. This is because, since April 2022, users need to be authenticated to access the data via S3. https://commoncrawl.org/access-the-data/.

Let me know what you think!